feat: refonte Settings→Setup par module, sessions, recorder vidéo & extensions moteurs
Regroupe le WIP accumulé (sources uniquement — zips prebuilt exclus, >100 Mo).

UI / Setup
- Suppression de SettingsWindow + settings/SystemSettingsTab au profit de
  panneaux Setup par module (Score, MidiScore, Timbre, Voice, VideoScroll,
  Sampler, LuxGrain, Source).
- AboutDialog, SamplerValueBox, param-chips éditables sampler, agrandissement
  polices UI, LookAndFeel/UITheme.

Sessions & persistance
- session/ : SessionManager, Sp3sImporter, Sp3ctraPaths (sessions par projet,
  migration one-shot, autosave ValueTree).

Vidéo
- recorder/ : VideoRecorder (.mov HEVC macOS + stub), VIDEO MIX / VideoMixer,
  Video Scroll invert luminance + fond de cadre.

Moteurs & audio
- Changement de sample rate à chaud (reinit SR des 4 moteurs).
- Worker threads LuxStral rebuild à chaud.
- Famille SCORE : enable persistant découplé du transport.
- Sampler : loop 3 toggles, LOAD image + rotation + scrub, MIDI targets étendus.
- Extensions luxgrain/luxstral/luxsynth/luxwave, staging, multithreading, MIDI.

+++ vst/source/ui/AudioMixPanel.h
@@ -43,9 +43,9 @@ public:
             { ModuleType::LuxWave,  "luxwaveVolume"  },
             { ModuleType::LuxGrain, "luxgrainVolume" },
         };
-        for (int i = 0; i < kNumEngines; ++i)
+        for (size_t i = 0; i < kNumEngines; ++i)
         {
-            auto& s = strips[(size_t) i];
+            auto& s = strips[i];
             s.type   = kEngines[i].t;
             s.colour = moduleColour(kEngines[i].t);
             s.name   = moduleDisplayName(kEngines[i].t)
